applied new policy to all runtimes with tests
